refactor: #132 CodeEditorModal 컴포넌트 분리
- CodeEditorModal: 모달 열림/닫힘 상태 관리, 게임 입력 비활성화 처리
- CodeEditorModalContent: Monaco 에디터와 협업 기능, 언어/테마 상태 관리
- CodeEditorControls: 언어 선택, 테마 선택, 연결 상태 표시 UI
